folio_project/requirements.txt
appnope==0.1.3
asgiref==3.5.2
astroid==2.11.6
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-env==0.2.1
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
tomlkit==0.11.0
traitlets==5.3.0
typing_extensions==4.3.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
